backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
celery==5.3.4

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# Docker
docker==7.0.0

# API Utils
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Rate Limiting
slowapi==0.1.9

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Monitoring
prometheus-client==0.19.0

# CORS
python-multipart==0.0.6